removing unnecessary interface, renaming file, adjusting imports.
